% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{attrition.bias.test}
\alias{attrition.bias.test}
\title{Function attrition.bias.test (Conduct attrition bias tests)}
\usage{
attrition.bias.test(data.source, variables = c("X", "Y"), no.waves = 3)
}
\arguments{
\item{data.source}{name of dataframe in wide format.}

\item{variables}{variables to be compared between response and non-response groups.}

\item{no.waves}{number of waves to be compared.}
}
\value{
Plot Response rates as p and attrition bias tests results.
}
\description{
Conduct attrition bias tests for lag = 1 and lag = 2.
}
\details{
For lag = 1, compares the means of a variable at T1 between the response and non-response groups at T2.
For lag = 2, compares the means of a variable at T1 between the response and non-response groups at T3.
Bonferroni adjustment for the number of variables (comparisons) at each time point is recommended.
}
\examples{

## -- Example -- ##

attrition.bias.test(data.source=Trust, variables=c("trust", "lonely", "lifesat"), no.waves=6)

}
