% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_loop.R
\name{c_loop}
\alias{c_loop}
\title{loop vector}
\usage{
c_loop(...)
}
\arguments{
\item{...}{vectors to be transformed into a list}
}
\description{
\code{c_loop()} takes in a set of vectors and transposes them into a
list of vectors. The length of the resulting list is the same as
each of the values passed to it. vectors of length 1 are recycled
throughout the results.
}
\examples{
c_loop(1:3,2:4,3:5,-1)

}
