% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP - Copy.R, R/LOOP.R
\name{CLPM}
\alias{CLPM}
\title{Function Cross-Lagged Panel Model (CLPM)}
\usage{
CLPM(data.source, no.waves, lag = 1, p = 0.001, X, Y, Z = "NULL", W = "NULL")

CLPM(data.source, no.waves, lag = 1, p = 0.001, X, Y, Z = "NULL", W = "NULL")
}
\arguments{
\item{data.source}{name of data.frame}

\item{no.waves}{number of waves (minimum = 3, must be grater than lag)}

\item{lag}{number of waves between two lags (minimum = 1, maximum = 4)}

\item{p}{critical p-value for pairwise comparisons (default is 0.001)}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
CLPM outputs.

CLPM outputs.
}
\description{
Cross-Lagged Panel Model

Cross-Lagged Panel Model
}
\examples{

## -- Example -- ##

CLPM(data.source="Data_A", 7, 2, X="EXPOSE", Y="INTENS")


## -- Example -- ##

CLPM(data.source="Data_A", 7, 2, X="EXPOSE", Y="INTENS")

}
