% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{LCS}
\alias{LCS}
\title{Function LCS (Latent Change Score Model)}
\usage{
LCS(
  data.source,
  no.waves,
  Type1 = 0.05,
  Type1Adj = "BON",
  X,
  Y,
  Z = "NULL",
  W = "NULL"
)
}
\arguments{
\item{data.source}{name of data.frame.}

\item{no.waves}{number of waves (minimum = 3).}

\item{Type1}{Overall Type I error rate (default is 0.05) for comparing estimated parameters in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each estimated parameter. Default is "BON" (Bonferroni adjustment -- Type I error rate/no. of pairwise comparisons), can also be "NULL" (without adjustment).}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
Latent Change Score Model (LCS) outputs.
}
\description{
Latent Change Score Model (LCS)
}
\examples{

## -- Example -- ##

LCS(data.source="Data_A", no.waves=6, Type1=0.05, Type1Adj="BON", X="EXPOSE.", Y="INTENS.")

}
