% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{LCSCC}
\alias{LCSCC}
\title{Function LCSCC (Latent Change Score Model with Changes-To-Changes Extension)}
\usage{
LCSCC(
  data.source,
  no.waves,
  varI.eq = FALSE,
  p = 0.001,
  X,
  Y,
  Z = "NULL",
  W = "NULL"
)
}
\arguments{
\item{data.source}{name of data.frame.}

\item{no.waves}{number of waves (minimum = 3).}

\item{varI.eq}{whether indicator residual variances are constrained to be equal (default is false).}

\item{p}{critical p-value for pairwise comparisons (default is 0.001).}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
Latent Change Score Model with Changes-To-Changes Extension (LCSCC) outputs.
}
\description{
Latent Change Score Model with Changes-To-Changes Extension (LCS-CC)
}
\examples{

## -- Example -- ##

LCSCC(data.source="Data_A", 7, varI.eq=TRUE, X="EXPOSE", Y="INTENS")

}
