% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{long2wide}
\alias{long2wide}
\title{Function long2wide (Convert Long Data Format to Wide Data Format)}
\usage{
long2wide(data.source, id = "id", time = "time", variables = c("X", "Y"))
}
\arguments{
\item{data.source}{name of dataframe.}

\item{id}{variable indicates group.}

\item{time}{variable indicates time.}
}
\value{
object as a dataframe.
}
\description{
Convert long data format to wide data format. New variables will be coded with ".1, .2" to indicate the timelines.
}
\examples{

## -- Example -- ##

df_wide <- long2wide(data.source=Responsive, id="id", time="time",
variables=c("act_con", "pas_con", "act_des", "pas_des", "engage", "satisfaction"))
write.csv(df_wide, "file_wide.csv") # save dataframe to csv file

}
