% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{loop on input}
\usage{
loop(fun, type = list)
}
\arguments{
\item{fun}{function to be looped over}

\item{type}{type of output, created via a call to \code{as._type_()}. A few special
tokens are permitted:
\itemize{
\item \code{df_rows} - returns a data.frame, where each result is a set of
rows
\item \code{mat_rows} - returns a matrix, where each result is a row
\item \code{mat_cols} - returns a matrix, where each result is a column
\item \code{simplify} - returns a simplified output, similar to \code{sapply()}
}}
}
\description{
\code{loop()} unifies \code{lapply} and \code{Map}. It supports a dollar notation so
that we can leverage the autompletion of function arguments.
Arguments to be looped on are prefixed with \code{+}.
}
\examples{
x <- loop(rnorm)(10,mean=+1:3)
x

loop(mean,numeric)(+x)

l <- list(iris, cars)
loop$head(+l, 1)
loop$head(+l, +1:2) # no need to worry about operator precedence

}
