% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looper.R
\name{looper}
\alias{looper}
\title{Loop over an expression}
\usage{
looper(..., type = list)
}
\arguments{
\item{...}{arguments to loop over, followed by a expression to perform}

\item{type}{type of output, created via a call to \code{as._type_()}. A few special
tokens are permitted:
\itemize{
\item \code{df_rows} - returns a data.frame, where each result is a set of
rows
\item \code{mat_rows} - returns a matrix, where each result is a row
\item \code{mat_cols} - returns a matrix, where each result is a column
\item \code{simplify} - returns a simplified output, similar to \code{sapply()}
}}
}
\description{
\code{looper()} is a wrapper around \code{loop()}. It takes a set of looping
arguments first, and then an expression to loop over as it's last
argument. This is closer in appearance to a \verb{for()} loop, but can
loop over multiple vectors.
}
\examples{
looper(n = +1:3,mean=+0:2,sd=2,
  type = df_rows,
  {
    data.frame(
      col_1 = rnorm(n),
      col_2 = rnorm(n,mean,sd)
      )
  })

}
