% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{ML}
\alias{ML}
\title{Function ML (Multilevel Model)}
\usage{
ML(
  model,
  data.source,
  Cluster = "NULL",
  missing = "listwise",
  L2 = TRUE,
  mL2.variables = NULL
)
}
\arguments{
\item{model}{User-specified measurement model.}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{missing}{missing = "listwise" or "fiml". lavaan uses listwise deletion for missing cases by default. missing = "fiml" in lavaan multilevel model is new and may have convergence problems.}

\item{L2}{If the model should be replicated at L2 (default = TRUE). If L2=FALSE, covariance among variables will be estimated at L2.}

\item{mL2.variables}{L2 variables represented by group mean (gm_X, manifest L2 variables) and L1 variables represented by group-mean centered variables (gmc_X). By default, variables are partitioned into L1 and L2 by multilevel SEM.}
}
\value{
Estimated parameters and differences in estimated parameters across levels
}
\description{
Multilevel Model allows for both MLM and MSEM and compares paths across levels
}
\details{
Define the model only once without specifying the level, and the function will compare the model across levels.
}
\examples{

## == Example == ##
# Data file is "Data_A"; cluster variable is "ID"

# Convert wide-format date file to long-format
# df_long <- wide2long(Data_A, no.waves=7, variables=c("EDU", "AGE", "BMI_S", "EXPOSE", "INTENS"), lag1=FALSE) 

## Specify the path model - Model ##
Model <- '
 INTENS ~ EDU + AGE + BMI_S
 EXPOSE ~ EDU + AGE + BMI_S
'

## ===== Compare Estimated Parameters Across Levels ===== ##
ML(Model, df_long, Cluster = "id", L2=TRUE, mL2.variables=c("EDU", "AGE", "BMI_S"))

}
