% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Responsive}
\alias{Responsive}
\title{Responsive dataset}
\format{
A data frame with 1620 rows and 13 variables:
\describe{
  \item{id}{Numeric, case id number.}
  \item{gender}{Numeric, Employee gender: 0 = female and 1 = male.}
  \item{age}{Numeric, Employee age.}
  \item{edu}{Numeric, Employee education level.}
  \item{tenure}{Numeric, Employee organizational tenure.}
  \item{proactive}{Numeric, Average of 4 proactive personality items: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{time}{Numeric, Day of data collection.}
  \item{act_con}{Numeric, Average of 3 active-constructive items: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{pas_con}{Numeric, Average of 3 passive-constructive items: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{act_des}{Numeric, Average of 3 active-destructive responsiveness items: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{pas_des}{Numeric, Average of 3 passive-destructive responsiveness items: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{engage}{Numeric, Average of 9 work engagement items: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{satisfaction}{Numeric, Average of 6 job satisfaction items: from 1 (strongly disagree) to 7 (strongly agree).}
}
}
\source{
Simulated dataset (180 observations) based on Liu, X., Dong, J., Yu, Y., Zheng, J., Zheng, X., & Lee, B. Y. (2026). From friction to fulfillment: Examining when and how spousal active-destructive responsiveness to employees' sharing of positive events benefits the work domain. Journal of Organizational Behavior, 47,  66-80.
}
\usage{
Responsive
}
\description{
Simulated daily diary dataset for demonstrating ML function.
}
\keyword{datasets}
