% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP - Copy.R, R/LOOP.R
\name{RICLPM}
\alias{RICLPM}
\title{Function Random Intercept Cross-Lagged Panel Model (RICLPM)}
\usage{
RICLPM(data.source, no.waves, lag = 1, p = 0.001, X, Y, Z = "NULL", W = "NULL")

RICLPM(data.source, no.waves, lag = 1, p = 0.001, X, Y, Z = "NULL", W = "NULL")
}
\arguments{
\item{data.source}{name of data.frame}

\item{no.waves}{number of waves (minimum = 3, must be grater than lag)}

\item{lag}{number of waves between two lags (minimum = 1, maximum = 4)}

\item{p}{critical p-value for pairwise comparisons (default is 0.001)}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
RICLPM outputs.

RICLPM outputs.
}
\description{
Random Intercept Cross-Lagged Panel Model (RI-CLPM)

Random Intercept Cross-Lagged Panel Model (RI-CLPM)
}
\examples{

## -- Example -- ##

RICLPM(data.source="Data_A", 7, 2, X="EXPOSE", Y="INTENS")


## -- Example -- ##

RICLPM(data.source="Data_A", 7, 2, X="EXPOSE", Y="INTENS")

}
