% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_loops.R
\name{simplified_loops}
\alias{simplified_loops}
\alias{loop_dbl}
\alias{loop_int}
\alias{loop_chr}
\alias{loop_lgl}
\alias{loop_dfr}
\alias{loop_mtr}
\alias{loop_mtc}
\alias{loop_smp}
\title{Simplifying Loops}
\usage{
loop_dbl(fun)

loop_int(fun)

loop_chr(fun)

loop_lgl(fun)

loop_dfr(fun)

loop_mtr(fun)

loop_mtc(fun)

loop_smp(fun)
}
\arguments{
\item{fun}{function to be looped over}
}
\description{
These functions are wrappers around \code{loop()} with a
pre-specified return type. This allows the \code{$} notation to be used
with a predictable outcome
}
\section{Functions}{
\itemize{
\item \code{loop_dbl}: outputs a double vector

\item \code{loop_int}: outputs an integer vector

\item \code{loop_chr}: outputs a character vector

\item \code{loop_lgl}: outputs a logical vector

\item \code{loop_dfr}: outputs a data.frame

\item \code{loop_mtr}: outputs a matrix joined by rows

\item \code{loop_mtc}: outputs a matrix joined by columns

\item \code{loop_smp}: outputs a \code{sapply()} style output
}}

\examples{
x <- loop$rnorm(n=100,mean=+1:20)
loop_dbl$mean(+x)
loop_int$length(+x)
loop_chr$paste0(+letters,+LETTERS)
loop_mtr$rnorm(n=10,mean=+5:8)
loop_mtc$rnorm(n=10,mean=+5:8)
loop_smp$rnorm(n=10,mean=+5:8)
}
