% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trust}
\alias{Trust}
\title{Trust dataset}
\format{
A data frame with 1000 rows and 19 variables:
\describe{
  \item{ID}{Numeric, case ID number.}
  \item{trust_T1}{Numeric, Average of 6 Trust items at Time 1: from 1 (strongly disagree) to 5 (strongly agree).}
  \item{trust_T2}{Numeric, Average of 6 Trust items at Time 2: from 1 (strongly disagree) to 5 (strongly agree).}
  \item{trust_T3}{Numeric, Average of 6 Trust items at Time 3: from 1 (strongly disagree) to 5 (strongly agree).}
  \item{trust_T4}{Numeric, Average of 6 Trust items at Time 4: from 1 (strongly disagree) to 5 (strongly agree).}
  \item{trust_T5}{Numeric, Average of 6 Trust items at Time 5: from 1 (strongly disagree) to 5 (strongly agree).}
  \item{trust_T6}{Numeric, Average of 6 Trust items at Time 6: from 1 (strongly disagree) to 5 (strongly agree).}
  \item{lonely_T1}{Numeric, Average of 6 UCLS Loneliness Scale at Time 1: from 1 (never) to 4 (always).}
  \item{lonely_T2}{Numeric, Average of 6 UCLS Loneliness Scale at Time 2: from 1 (never) to 4 (always).}
  \item{lonely_T3}{Numeric, Average of 6 UCLS Loneliness Scale at Time 3: from 1 (never) to 4 (always).}
  \item{lonely_T4}{Numeric, Average of 6 UCLS Loneliness Scale at Time 4: from 1 (never) to 4 (always).}
  \item{lonely_T5}{Numeric, Average of 6 UCLS Loneliness Scale at Time 5: from 1 (never) to 4 (always).}
  \item{lonely_T6}{Numeric, Average of 6 UCLS Loneliness Scale at Time 6: from 1 (never) to 4 (always).}
  \item{lifesat_T1}{Numeric, Average of 5 Satisfaction with Life Scale at Time 1: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{lifesat_T2}{Numeric, Average of 5 Satisfaction with Life Scale at Time 2: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{lifesat_T3}{Numeric, Average of 5 Satisfaction with Life Scale at Time 3: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{lifesat_T4}{Numeric, Average of 5 Satisfaction with Life Scale at Time 4: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{lifesat_T5}{Numeric, Average of 5 Satisfaction with Life Scale at Time 5: from 1 (strongly disagree) to 7 (strongly agree).}
  \item{lifesat_T6}{Numeric, Average of 5 Satisfaction with Life Scale at Time 6: from 1 (strongly disagree) to 7 (strongly agree).}
}
}
\source{
Simulated dataset (500 observations) based on Chen, Y., Fang, Y., Yang, Y., & Dong, Y. (2026) Longitudinal associations among trust, loneliness, and life satisfaction among university students: A between- and within-person analysis.Journal of Personality, 94: 151-162.
}
\usage{
Trust
}
\description{
Simulated dataset for demonstrating LOOP functions.
}
\keyword{datasets}
