% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{create.lag}
\alias{create.lag}
\title{Function create.lag (Create the lag variable by 1 time point)}
\usage{
create.lag(data.source, id = "id", time = "time", variables = c("X", "Y"))
}
\arguments{
\item{data.source}{name of dataframe in long format with a grouping and a timing variable.}

\item{id}{group identification variable.}

\item{time}{timing variable.}

\item{variables}{variable names to create lagged (t-1) variables.}
}
\value{
dataframe df_lagged.
}
\description{
Create lagged (t-1) variables in dataframe.
}
\examples{

## -- Example -- ##


df_lagged <- create.lag(data.source=Responsive, id="id", time="time", 
variables=c("satisfaction", "engage"))
write.csv(df_lagged, "Responsive_lagged.csv") # save dataframe to csv file

}
