% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{GCLM}
\alias{GCLM}
\title{Function GCLM (General Cross-Lagged Panel Model )}
\usage{
GCLM(
  data.source,
  no.waves,
  lag = 1,
  Type1 = 0.05,
  Type1Adj = "BON",
  X,
  Y,
  Z = "NULL",
  W = "NULL"
)
}
\arguments{
\item{data.source}{name of data.frame}

\item{no.waves}{number of waves (minimum = 3, must be grater than AR & MA)}

\item{lag}{number of lags in autoregressive effect (AR) and moving average (MA) (default = 1) AR = MA}

\item{Type1}{Overall Type I error rate (default is 0.05) for comparing estimated parameters in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each estimated parameter. Default is "BON" (Bonferroni adjustment -- Type I error rate/no. of pairwise comparisons), can also be "NULL" (without adjustment).}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
General Cross-Lagged Panel Model (GCLM) outputs.
}
\description{
General Cross-Lagged Panel Model (GCLM)
}
\examples{

## -- Example -- ##

GCLM(data.source="Smoking", no.waves=6, lag=1, Type1=0.05, Type1Adj="BON", X="expose.", 
Y="intens.")

}
