% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{GCLM}
\alias{GCLM}
\title{Function GCLM (General Cross-Lagged Panel Model )}
\usage{
GCLM(data.source, no.waves, lag = 1, p = 0.001, X, Y, Z = "NULL", W = "NULL")
}
\arguments{
\item{data.source}{name of data.frame}

\item{no.waves}{number of waves (minimum = 3, must be grater than AR & MA)}

\item{lag}{number of lags in autoregressive effect (AR) and moving average (MA) (default = 1) AR = MA}

\item{p}{critical p-value for pairwise comparisons (default is 0.001)}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
General Cross-Lagged Panel Model (GCLM) outputs.
}
\description{
General Cross-Lagged Panel Model (GCLM)
}
\examples{

## -- Example -- ##

GCLM(data.source="Data_A", 7, lag=1, X="EXPOSE", Y="INTENS")

}
