% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{LCS}
\alias{LCS}
\title{Function LCS (Latent Change Score Model)}
\usage{
LCS(data.source, no.waves, p = 0.001, X, Y, Z = "NULL", W = "NULL")
}
\arguments{
\item{data.source}{name of data.frame.}

\item{no.waves}{number of waves (minimum = 3).}

\item{p}{critical p-value for pairwise comparisons (default is 0.001).}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
Latent Change Score Model (LCS) outputs.
}
\description{
Latent Change Score Model (LCS)
}
\examples{

## -- Example -- ##

LCS(data.source="Data_A", 7, X="EXPOSE.", Y="INTENS.")

}
