% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{LCSCC}
\alias{LCSCC}
\title{Function LCSCC (Latent Change Score Model with Changes-To-Changes Extension)}
\usage{
LCSCC(
  data.source,
  no.waves,
  varI.eq = FALSE,
  Type1 = 0.05,
  Type1Adj = "BON",
  X,
  Y,
  Z = "NULL",
  W = "NULL"
)
}
\arguments{
\item{data.source}{name of data.frame.}

\item{no.waves}{number of waves (minimum = 3).}

\item{varI.eq}{whether indicator residual variances are constrained to be equal (default is false).}

\item{Type1}{Overall Type I error rate (default is 0.05) for comparing estimated parameters in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each estimated parameter. Default is "BON" (Bonferroni adjustment -- Type I error rate/no. of pairwise comparisons), can also be "NULL" (without adjustment).}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
Latent Change Score Model with Changes-To-Changes Extension (LCSCC) outputs.
}
\description{
Latent Change Score Model with Changes-To-Changes Extension (LCS-CC)
}
\examples{

## -- Example -- ##

LCSCC(data.source="Data_A", no.waves=6, varI.eq=TRUE, Type1=0.05, Type1Adj="BON", X="EXPOSE.", 
Y="INTENS.")

}
