% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{LGCM}
\alias{LGCM}
\title{Function Latent Growth Curve Model (LGCM)}
\usage{
LGCM(data.source, no.waves, p = 0.001, X, Y, Z = "NULL", W = "NULL")
}
\arguments{
\item{data.source}{name of data.frame}

\item{no.waves}{number of waves (minimum = 3)}

\item{p}{critical p-value for pairwise comparisons (default is 0.001)}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
LGCM outputs.
}
\description{
Latent Growth Curve Model (LGCM)
}
\examples{

## -- Example -- ##

LGCM(data.source="Data_A", 7, X="EXPOSE", Y="INTENS")

}
