% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loop}
\alias{loop}
\alias{loop2}
\title{loop on input}
\usage{
loop(fun)

loop2(fun)
}
\arguments{
\item{fun}{}
}
\value{

}
\description{
\code{loop} unifies \code{lapply} and \code{Map}, \code{loop2} unifies \code{sapply} and \code{mapply}.
Both support a dollar notation so that we can leverage the autompletion of
function arguments.
}
\examples{
l <- list(iris, cars)
loop$head(+l, 1)
loop$head(+l, +1:2)
loop2$nrow()
}
