% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{RICLPM}
\alias{RICLPM}
\title{Function RICLPM (Random Intercept Cross-Lagged Panel Model)}
\usage{
RICLPM(
  data.source,
  no.waves,
  lag = 1,
  Type1 = 0.05,
  Type1Adj = "BON",
  X,
  Y,
  Z = "NULL",
  W = "NULL"
)
}
\arguments{
\item{data.source}{name of data.frame}

\item{no.waves}{number of waves (minimum = 3, must be grater than lag)}

\item{lag}{number of waves between two lags (minimum = 1, maximum = 4)}

\item{Type1}{Overall Type I error rate (default is 0.05) for comparing estimated parameters in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each estimated parameter. Default is "BON" (Bonferroni adjustment -- Type I error rate/no. of pairwise comparisons), can also be "NULL" (without adjustment).}

\item{X}{name of variable X.}

\item{Y}{name of variable Y.}

\item{Z}{name of variable Z.}

\item{W}{name of variable W (Z must come before W).}
}
\value{
Random Intercept Cross-Lagged Panel Model (RI-CLPM) outputs.
}
\description{
Random Intercept Cross-Lagged Panel Model (RI-CLPM)
}
\examples{

## -- Example -- ##

RICLPM(data.source="Data_A", no.waves=6, lag=2, Type1=0.05, Type1Adj="BON", X="EXPOSE.", 
Y="INTENS.")

}
