% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Smoking}
\alias{Smoking}
\title{Smoking dataset}
\format{
A data frame with 500 rows and 13 variables:
\describe{
  \item{id}{Numeric, case id number.}
  \item{expose.1}{Numeric, Perceived exposure to smoking in movies at Time 1.}
  \item{expose.2}{Numeric, Perceived exposure to smoking in movies at Time 2.}
  \item{expose.3}{Numeric, Perceived exposure to smoking in movies at Time 3.}
  \item{expose.4}{Numeric, Perceived exposure to smoking in movies at Time 4.}
  \item{expose.5}{Numeric, Perceived exposure to smoking in movies at Time 5.}
  \item{expose.6}{Numeric, Perceived exposure to smoking in movies at Time 6.}
  \item{intens.1}{Numeric, Smoking intensity at Time 1.}
  \item{intens.2}{Numeric, Smoking intensity at Time 2.}
  \item{intens.3}{Numeric, Smoking intensity at Time 3.}
  \item{intens.4}{Numeric, Smoking intensity at Time 4.}
  \item{intens.5}{Numeric, Smoking intensity at Time 5.}
  \item{intens.6}{Numeric, Smoking intensity at Time 6.}
}
}
\source{
Dataset with 500 observations across six time points simulated with lavaan::simulateData() based on results of the Minnesota Adolescent Community Cohort (MACC) Study, 2000-2013, reported in Usami, S., Murayama, K., & Hamaker, E. L. (2019). A unified framework of longitudinal models to examine reciprocal relations. Psychological Methods, 24(5), 637-657.
}
\usage{
Smoking
}
\description{
Simulated dataset for demonstrating GCLM() and STARTS() in longitudinal studies.
}
\keyword{datasets}
