% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Smoking}
\alias{Smoking}
\title{Smoking dataset}
\format{
A data frame with 500 rows and 22 variables:
\describe{
  \item{ID}{Numeric, case ID number.}
  \item{expose.1}{Numeric, Perceived exposure to smoking in movies at Time 1 from 1 (never) to 4 (most of the time).}
  \item{expose.2}{Numeric, Perceived exposure to smoking in movies at Time 2 from 1 (never) to 4 (most of the time).}
  \item{expose.3}{Numeric, Perceived exposure to smoking in movies at Time 3 from 1 (never) to 4 (most of the time).}
  \item{expose.4}{Numeric, Perceived exposure to smoking in movies at Time 4 from 1 (never) to 4 (most of the time).}
  \item{expose.5}{Numeric, Perceived exposure to smoking in movies at Time 5 from 1 (never) to 4 (most of the time).}
  \item{expose.6}{Numeric, Perceived exposure to smoking in movies at Time 6 from 1 (never) to 4 (most of the time).}
  \item{intens.1}{Numeric, Smoking intensity at Time 1 from 1 (non-current smoker) to 6 (weekly, heavy).}
  \item{intens.2}{Numeric, Smoking intensity at Time 2 from 1 (non-current smoker) to 6 (weekly, heavy).}
  \item{intens.3}{Numeric, Smoking intensity at Time 3 from 1 (non-current smoker) to 6 (weekly, heavy).}
  \item{intens.4}{Numeric, Smoking intensity at Time 4 from 1 (non-current smoker) to 6 (weekly, heavy).}
  \item{intens.5}{Numeric, Smoking intensity at Time 5 from 1 (non-current smoker) to 6 (weekly, heavy).}
  \item{intens.6}{Numeric, Smoking intensity at Time 6 from 1 (non-current smoker) to 6 (weekly, heavy).}
  \item{stage.1}{Numeric, Smoke stages at Time 1 from 1 (never) to 7 (quitter).}
  \item{stage.2}{Numeric, Smoke stages at Time 2 from 1 (never) to 7 (quitter).}
  \item{stage.3}{Numeric, Smoke stages at Time 3 from 1 (never) to 7 (quitter).}
  \item{stage.4}{Numeric, Smoke stages at Time 4 from 1 (never) to 7 (quitter).}
  \item{stage.5}{Numeric, Smoke stages at Time 5 from 1 (never) to 7 (quitter).}
  \item{stage.6}{Numeric, Smoke stages at Time 6 from 1 (never) to 7 (quitter).}
}
}
\source{
Simulated dataset with 500 observations across seven time points based on data from the Minnesota Adolescent Community Cohort (MACC) Study, 2000-2013, used in Usami, S., Murayama, K., & Hamaker, E. L. (2019). A unified framework of longitudinal models to examine reciprocal relations. Psychological Methods, 24(5), 637-657.
}
\usage{
Smoking
}
\description{
Simulated dataset for demonstrating GCLM() and STARTS() in longitudinal studies.
}
\keyword{datasets}
