% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{wide2long}
\alias{wide2long}
\title{Function wide2long (Convert Wide Data Format to Long Data Format)}
\usage{
wide2long(data.source, no.waves, variables = c("X", "Y"), lag1 = TRUE)
}
\arguments{
\item{data.source}{name of dataframe.}

\item{no.waves}{number of waves of data.}

\item{variables}{variable names to create long data format.}

\item{lag1}{whether to create lagged (t - 1) data of the variables. Default is FALSE.}
}
\value{
object as a dataframe.
}
\description{
Convert wide data format to long data format with the capability to create lagged variables. Variables should be coded with ".1, .2" to indicate the time frame. A new "id" variable will be created as the grouping variable, and a new "time" variable will be created as the timing variable.
}
\examples{

## -- Example -- ##

Data_B <- Data_A[1:50,]

# Suppose df_long is the new dataframe name.
# Suppese data are "EXPOSE.1", "EXPOSE.2", "EXPOSE.3", ... "EXPOSE.7" in Data_7

df_long <- wide2long(Data_B, 7, variables=c("EXPOSE", "INTENS"), lag1=TRUE) 

}
