% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOP.R
\name{wide2long}
\alias{wide2long}
\title{Function wide2long (Convert Wide Data Format to Long Data Format)}
\usage{
wide2long(data.source, no.waves, variables = c("X", "Y"), lag1 = FALSE)
}
\arguments{
\item{data.source}{name of dataframe.}

\item{no.waves}{number of waves of data.}

\item{variables}{variable names to create long data format.}

\item{lag1}{whether to create lagged (t - 1) data of the variables. Default is FALSE.}
}
\value{
object as a dataframe.
}
\description{
Convert wide data format to long data format with the capability to create lagged (t-1) variables. Variables should be coded with ".1, .2" to indicate the timelines. A new "id" variable will be created as the grouping variable, and a new "time" variable will be created as the timing variable.
}
\examples{

## -- Example -- ##


df_long <- wide2long(data.source=Trust, no.waves=6, variables=c("trust.", "lonely.", 
"lifesat."), lag1=FALSE)
write.csv(df_long, "file_long.csv") # save dataframe to csv file

}
